Issue #6406 - move JaspiConfiguration into jetty-jaspi module
