+++ jetty-jaspi/src/main/resources/META-INF/services/org.eclipse.jetty.webapp.Configuration
@@ -1,0 +1,1 @@
+org.eclipse.jetty.security.jaspi.JaspiConfiguration
